Продолжаем

+++ mainService/src/main/java/ewm/API/publicAPI/events/EventPublicController.java
@@ -18,7 +18,7 @@ public class EventPublicController {
 
     @GetMapping
     public List<EventResponseDto> getAllEvents(@RequestParam(required = false) String text,
-                                               @RequestParam(required = false) List<Integer> categories,
+                                               @RequestParam(required = false) List<Long> categories,
                                                @RequestParam(required = false) Boolean paid,
                                                @RequestParam(required = false) @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss") LocalDateTime rangeStart,
                                                @RequestParam(required = false) @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss") LocalDateTime rangeEnd,
